+++ src/Core/Sektor.DarkestDungeon.Core.Content/Database/JsonHeirloomExchangeEntry.cs
@@ -1,5 +1,3 @@
-using Newtonsoft.Json;
-
 namespace Sektor.DarkestDungeon.Core.Content.Database
 {
     /// <summary>
@@ -8,19 +6,15 @@ namespace Sektor.DarkestDungeon.Core.Content.Database
     public class JsonHeirloomExchangeEntry
     {
         /// <summary>Gets the currency type being exchanged.</summary>
-        [JsonProperty("exchange_from_type")]
-        public string ExchangeFromType { get; set; }
+        public string exchange_from_type { get; set; }
 
         /// <summary>Gets the amount of the source currency paid for the exchange.</summary>
-        [JsonProperty("exchange_from_amount")]
-        public int ExchangeFromAmount { get; set; }
+        public int exchange_from_amount { get; set; }
 
         /// <summary>Gets the currency type received in exchange.</summary>
-        [JsonProperty("exchange_to_type")]
-        public string ExchangeToType { get; set; }
+        public string exchange_to_type { get; set; }
 
         /// <summary>Gets the amount of the target currency received in exchange.</summary>
-        [JsonProperty("exchange_to_amount")]
-        public int ExchangeToAmount { get; set; }
+        public int exchange_to_amount { get; set; }
     }
 }