+++ src/Core/Sektor.DarkestDungeon.Core.Content/Database/JsonPartyNameEntry.cs
@@ -1,7 +1,5 @@
 using System.Collections.Generic;
 
-using Newtonsoft.Json;
-
 namespace Sektor.DarkestDungeon.Core.Content.Database
 {
     /// <summary>
@@ -10,11 +8,9 @@ namespace Sektor.DarkestDungeon.Core.Content.Database
     public class JsonPartyNameEntry
     {
         /// <summary>Gets the unique identifier of the party name.</summary>
-        [JsonProperty("id")]
-        public string Id { get; set; }
+        public string id { get; set; }
 
         /// <summary>Gets the hero classes required for the party name to be available.</summary>
-        [JsonProperty("required_hero_class")]
-        public List<string> RequiredHeroClass { get; set; }
+        public List<string> required_hero_class { get; set; }
     }
 }